chore: refactor knuu to an object
