+++ pkg/knuu/knuu.go
@@ -18,184 +18,134 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 
 	"github.com/celestiaorg/knuu/pkg/builder"
-	"github.com/celestiaorg/knuu/pkg/builder/docker"
 	"github.com/celestiaorg/knuu/pkg/builder/kaniko"
+	"github.com/celestiaorg/knuu/pkg/instance"
 	"github.com/celestiaorg/knuu/pkg/k8s"
 	"github.com/celestiaorg/knuu/pkg/minio"
+	"github.com/celestiaorg/knuu/pkg/system"
 )
 
-var (
-	// testScope is the testScope of the current knuu instance
-	testScope    string
-	startTime    string
-	timeout      time.Duration
-	imageBuilder builder.Builder
-
-	// TODO: this temporary until we refactor knuu pkg
-	k8sClient *k8s.Client
+const (
+	defaultTimeout     = 60 * time.Minute
+	timeoutHandlerName = "timeout-handler"
+	// FIXME: use supported kubernetes version images (use of latest could break) (https://github.com/celestiaorg/knuu/issues/116)
+	timeoutHandlerImage = "docker.io/bitnami/kubectl:latest"
 )
 
-// Initialize initializes knuu with a unique scope
-func Initialize() error {
-	t := time.Now()
-	scope := fmt.Sprintf("%s-%03d", t.Format("20060102-150405"), t.Nanosecond()/1e6)
-	return InitializeWithScope(scope)
+type Knuu struct {
+	system.SystemDependencies
+	timeout time.Duration
 }
 
-// Scope returns the scope of the current knuu instance
-func Scope() string {
-	return testScope
+type Option func(*Knuu)
+
+func WithImageBuilder(builder builder.Builder) Option {
+	return func(k *Knuu) {
+		k.ImageBuilder = builder
+	}
 }
 
-// InitializeWithScope initializes knuu with a given scope
-func InitializeWithScope(scope string) error {
-	var err error
-	err = godotenv.Load()
-	if err != nil {
-		if os.IsNotExist(err) {
-			logrus.Info("The .env file does not exist, continuing without loading environment variables.")
-		} else {
-			return ErrCannotLoadEnv.Wrap(err)
+func WithTestScope(scope string) Option {
+	return func(k *Knuu) {
+		k.TestScope = k8s.SanitizeName(scope)
+	}
+}
+
+// This timeout indicates how long the test will run before it is considered failed.
+func WithTimeout(timeout time.Duration) Option {
+	return func(k *Knuu) {
+		k.timeout = timeout
+	}
+}
+
+func WithMinio(minio *minio.Minio) Option {
+	return func(k *Knuu) {
+		k.MinioCli = minio
+	}
+}
+
+func WithK8s(k8s *k8s.Client) Option {
+	return func(k *Knuu) {
+		k.K8sCli = k8s
+	}
+}
+
+func WithLogger(logger *logrus.Logger) Option {
+	return func(k *Knuu) {
+		k.Logger = logger
+	}
+}
+
+func New(ctx context.Context, opts ...Option) (*Knuu, error) {
+	if err := godotenv.Load(); err != nil {
+		if !os.IsNotExist(err) {
+			return nil, ErrCannotLoadEnv.Wrap(err)
 		}
-	}
-	if scope == "" {
-		return ErrCannotInitializeKnuuWithEmptyScope
+		logrus.Info("The .env file does not exist, continuing without loading environment variables.")
 	}
 
-	// Sanitize the scope to ensure it is a valid Kubernetes name and match the namespace
-	testScope = k8s.SanitizeName(scope)
-
-	t := time.Now()
-	startTime = fmt.Sprintf("%s-%03d", t.Format("20060102-150405"), t.Nanosecond()/1e6)
-
-	setupLogging()
-
-	// Override scope if KNUU_NAMESPACE is set
-	namespaceEnv := os.Getenv("KNUU_NAMESPACE")
-	if namespaceEnv != "" {
-		scope = namespaceEnv
-		logrus.Warnf("KNUU_NAMESPACE is deprecated. Scope overridden to: %s", scope)
+	k := &Knuu{}
+	for _, opt := range opts {
+		opt(k)
 	}
 
-	logrus.Infof("Initializing knuu with scope: %s", testScope)
+	k.StartTime = time.Now().UTC().Format("20060102T150405Z")
 
-	// read timeout from env
-	timeoutString := os.Getenv("KNUU_TIMEOUT")
-	if timeoutString == "" {
-		timeout = 60 * time.Minute
-	} else {
-		parsedTimeout, err := time.ParseDuration(timeoutString)
+	// handle default values
+	if k.Logger == nil {
+		k.Logger = defaultLogger()
+	}
+
+	if k.TestScope == "" {
+		t := time.Now()
+		k.TestScope = fmt.Sprintf("%s-%03d", t.Format("20060102-150405"), t.Nanosecond()/1e6)
+	}
+
+	if k.timeout == 0 {
+		k.timeout = defaultTimeout
+	}
+
+	if k.K8sCli == nil {
+		var err error
+		k.K8sCli, err = k8s.New(ctx, k.TestScope)
 		if err != nil {
-			return ErrCannotParseTimeout.Wrap(err)
+			return nil, ErrCannotInitializeK8s.Wrap(err)
 		}
-		timeout = parsedTimeout
 	}
 
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-
-	k8sClient, err = k8s.New(ctx, scope)
-	if err != nil {
-		return ErrCannotInitializeK8s.Wrap(err)
+	if k.MinioCli == nil {
+		// TODO: minio also needs a little refactor to accept k8s obj instead
+		k.MinioCli = &minio.Minio{
+			Clientset: k.K8sCli.Clientset(),
+			Namespace: k.K8sCli.Namespace(),
+		}
 	}
 
-	if err := handleTimeout(); err != nil {
-		return ErrCannotHandleTimeout.Wrap(err)
+	if k.ImageBuilder == nil {
+		// TODO: Also here for kaniko
+		k.ImageBuilder = &kaniko.Kaniko{
+			K8sClientset: k.K8sCli.Clientset(),
+			K8sNamespace: k.K8sCli.Namespace(),
+			Minio:        k.MinioCli,
+		}
 	}
 
-	minioClient = &minio.Minio{
-		Clientset: k8sClient.Clientset(),
-		Namespace: k8sClient.Namespace(),
+	if err := k.handleTimeout(ctx); err != nil {
+		return nil, ErrCannotHandleTimeout.Wrap(err)
 	}
 
-	builderType := os.Getenv("KNUU_BUILDER")
-	switch builderType {
-	case "kubernetes":
-		SetImageBuilder(&kaniko.Kaniko{
-			K8sClientset: k8sClient.Clientset(),
-			K8sNamespace: k8sClient.Namespace(),
-			Minio:        minioClient, // same client is used to make the best use of the resources
-		})
-	case "docker", "":
-		SetImageBuilder(&docker.Docker{
-			K8sClientset: k8sClient.Clientset(),
-			K8sNamespace: k8sClient.Namespace(),
-		})
-	default:
-		return ErrInvalidKnuuBuilder.WithParams(builderType)
-	}
-
-	HandleStopSignal()
-	return nil
+	return k, nil
 }
 
-// Deprecated: Identifier is deprecated, use Scope() instead.
-func Identifier() string {
-	logrus.Warn("Identifier() is deprecated, use Scope() instead.")
-	return Scope()
+func (k *Knuu) Scope() string {
+	return k.TestScope
 }
 
-// Deprecated: InitializeWithIdentifier is deprecated, use InitializeWithScope(scope string) instead.
-func InitializeWithIdentifier(uniqueIdentifier string) error {
-	logrus.Warn("InitializeWithIdentifier is deprecated, use InitializeWithScope(scope string) instead.")
-	return InitializeWithScope(uniqueIdentifier)
+func (k *Knuu) CleanUp(ctx context.Context) error {
+	return k.K8sCli.DeleteNamespace(ctx, k.TestScope)
 }
 
-// setupLogging Configures the log
-func setupLogging() {
-	// Set the default log level
-	logrus.SetLevel(logrus.InfoLevel)
-
-	// Set the custom formatter
-	logrus.SetFormatter(&logrus.TextFormatter{
-		FullTimestamp: true,
-		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
-			filename := path.Base(f.File)
-			directory := path.Base(path.Dir(f.File))
-			return "", directory + "/" + filename + ":" + strconv.Itoa(f.Line)
-		},
-	})
-
-	// Enable reporting the file and line
-	logrus.SetReportCaller(true)
-
-	switch os.Getenv("LOG_LEVEL") {
-	case "debug":
-		logrus.SetLevel(logrus.DebugLevel)
-	case "info":
-		logrus.SetLevel(logrus.InfoLevel)
-	case "warn":
-		logrus.SetLevel(logrus.WarnLevel)
-	case "error":
-		logrus.SetLevel(logrus.ErrorLevel)
-	default:
-		logrus.SetLevel(logrus.InfoLevel)
-	}
-
-	logrus.Info("LOG_LEVEL: ", logrus.GetLevel())
-}
-
-func SetImageBuilder(b builder.Builder) {
-	imageBuilder = b
-}
-
-func ImageBuilder() builder.Builder {
-	return imageBuilder
-}
-
-// IsInitialized returns true if knuu is initialized, and false otherwise
-func IsInitialized() bool {
-	return k8sClient != nil
-}
-
-func CleanUp() error {
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-
-	return k8sClient.DeleteNamespace(ctx, testScope)
-}
-
-func HandleStopSignal() {
+func (k *Knuu) HandleStopSignal() {
 	stop := make(chan os.Signal, 1)
 	signal.Notify(stop, syscall.SIGINT, syscall.SIGTERM, os.Interrupt)
 	go func() {
@@ -208,17 +158,17 @@ func HandleStopSignal() {
 }
 
 // handleTimeout creates a timeout handler that will delete all resources with the scope after the timeout
-func handleTimeout() error {
-	instance, err := NewInstance("timeout-handler")
+func (k *Knuu) handleTimeout(ctx context.Context) error {
+	inst, err := k.NewInstance(timeoutHandlerName)
 	if err != nil {
 		return ErrCannotCreateInstance.Wrap(err)
 	}
-	instance.instanceType = TimeoutHandlerInstance
-	// FIXME: use supported kubernetes version images (use of latest could break) (https://github.com/celestiaorg/knuu/issues/116)
-	if err := instance.SetImage("docker.io/bitnami/kubectl:latest"); err != nil {
+	inst.SetInstanceType(instance.TimeoutHandlerInstance)
+
+	if err := inst.SetImage(ctx, timeoutHandlerImage); err != nil {
 		return ErrCannotSetImage.Wrap(err)
 	}
-	if err := instance.Commit(); err != nil {
+	if err := inst.Commit(); err != nil {
 		return ErrCannotCommitInstance.Wrap(err)
 	}
 
@@ -226,24 +176,26 @@ func handleTimeout() error {
 
 	// Wait for a specific period before executing the next operation.
 	// This is useful to ensure that any previous operation has time to complete.
-	commands = append(commands, fmt.Sprintf("sleep %d", int64(timeout.Seconds())))
+	commands = append(commands, fmt.Sprintf("sleep %d", int64(k.timeout.Seconds())))
 	// Collects all resources (pods, services, etc.) within the specified namespace that match a specific label, excluding certain types,
 	// and then deletes them. This is useful for cleaning up specific test resources before proceeding to delete the namespace.
-	commands = append(commands, fmt.Sprintf("kubectl get all,pvc,netpol,roles,serviceaccounts,rolebindings,configmaps -l knuu.sh/scope=%s -n %s -o json | jq -r '.items[] | select(.metadata.labels.\"knuu.sh/type\" != \"%s\") | \"\\(.kind)/\\(.metadata.name)\"' | xargs -r kubectl delete -n %s", testScope, k8sClient.Namespace(), TimeoutHandlerInstance.String(), k8sClient.Namespace()))
+	commands = append(commands,
+		fmt.Sprintf("kubectl get all,pvc,netpol,roles,serviceaccounts,rolebindings,configmaps -l knuu.sh/scope=%s -n %s -o json | jq -r '.items[] | select(.metadata.labels.\"knuu.sh/type\" != \"%s\") | \"\\(.kind)/\\(.metadata.name)\"' | xargs -r kubectl delete -n %s",
+			k.TestScope, k.K8sCli.Namespace(), instance.TimeoutHandlerInstance.String(), k.K8sCli.Namespace()))
 
 	// Delete the namespace as it was created by knuu.
-	logrus.Debugf("The namespace generated [%s] will be deleted", k8sClient.Namespace())
-	commands = append(commands, fmt.Sprintf("kubectl delete namespace %s", k8sClient.Namespace()))
+	k.Logger.Debugf("The namespace generated [%s] will be deleted", k.K8sCli.Namespace())
+	commands = append(commands, fmt.Sprintf("kubectl delete namespace %s", k.K8sCli.Namespace()))
 
 	// Delete all labeled resources within the namespace.
 	// Unlike the previous command that excludes certain types, this command ensures that everything remaining is deleted.
-	commands = append(commands, fmt.Sprintf("kubectl delete all,pvc,netpol,roles,serviceaccounts,rolebindings,configmaps -l knuu.sh/scope=%s -n %s", testScope, k8sClient.Namespace()))
+	commands = append(commands, fmt.Sprintf("kubectl delete all,pvc,netpol,roles,serviceaccounts,rolebindings,configmaps -l knuu.sh/scope=%s -n %s", k.TestScope, k.K8sCli.Namespace()))
 
 	finalCmd := strings.Join(commands, " && ")
 
 	// Run the command
-	if err := instance.SetCommand("sh", "-c", finalCmd); err != nil {
-		logrus.Debugf("The full command generated is [%s]", finalCmd)
+	if err := inst.SetCommand("sh", "-c", finalCmd); err != nil {
+		k.Logger.Debugf("The full command generated is [%s]", finalCmd)
 		return ErrCannotSetCommand.Wrap(err)
 	}
 
@@ -253,12 +205,39 @@ func handleTimeout() error {
 		Resources: []string{"*"},
 	}
 
-	if err := instance.AddPolicyRule(rule); err != nil {
+	if err := inst.AddPolicyRule(rule); err != nil {
 		return ErrCannotAddPolicyRule.Wrap(err)
 	}
-	if err := instance.Start(); err != nil {
+	if err := inst.Start(ctx); err != nil {
 		return ErrCannotStartInstance.Wrap(err)
 	}
 
 	return nil
+}
+
+func defaultLogger() *logrus.Logger {
+	logger := logrus.New()
+
+	logger.SetFormatter(&logrus.TextFormatter{
+		FullTimestamp: true,
+		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
+			filename := path.Base(f.File)
+			directory := path.Base(path.Dir(f.File))
+			return "", directory + "/" + filename + ":" + strconv.Itoa(f.Line)
+		},
+	})
+
+	// Enable reporting the file and line
+	logger.SetReportCaller(true)
+
+	customLevel := os.Getenv("LOG_LEVEL")
+	if customLevel != "" {
+		err := logger.Level.UnmarshalText([]byte(customLevel))
+		if err != nil {
+			logger.Warnf("Failed to parse LOG_LEVEL: %v, defaulting to INFO", err)
+		}
+	}
+	logger.Info("LOG_LEVEL: ", logger.GetLevel())
+
+	return logger
 }
